requirements.txt
markdown>=3.3.0
numpy>=1.16.0
jinja2>=3.0
Pygments>=2.9.0
flask>=2.0.0
flask-socketio>=5.1.0
pymdown-extensions>=8.0
flask-cors>=3.0.0
Pillow>=8.0
qrcode>=7.3
